+++ lib/src/http_cache.dart
@@ -5,10 +5,10 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:http_cache_flutter/src/debug_configuration.dart';
 import 'package:http_cache_flutter/src/error_impl.dart';
+import 'package:http_cache_flutter/src/hc_log.dart';
 import 'package:http_cache_flutter/src/http_cache_builder_data.dart';
 import 'package:http_cache_flutter/src/http_cache_chiper.dart';
 import 'package:http_cache_flutter/src/http_cache_storage.dart';
-import 'dart:developer' as developer;
 
 import 'package:http/http.dart' as http;
 
@@ -20,7 +20,7 @@ class HttpCache<T> extends StatefulWidget {
   ///your header request, this data provided by http package.
   final Map<String, String>? headers;
 
-  final Future<Map<String, dynamic>>? futureHeaders;
+  final Future<Map<String, String>>? futureHeaders;
 
   ///this callback will run when the fetch got an error.
   final Function(Object error)? onError;
@@ -105,24 +105,14 @@ class _HttpCacheState<T> extends State<HttpCache<T>> {
     }
 
     response = HttpResponse.fromMap(data);
-    if (widget.log.showLog) {
-      developer.log(
-        response?.statusCode.toString() ?? "",
-        name: '[local response] Status code',
-        level: widget.log.level,
-      );
-      developer.log(
-        response?.headers.toString() ?? "",
-        name: '[local response] header',
-        level: widget.log.level,
-      );
-      developer.log(
-        response?.body ?? "",
-        name: '[local response] body',
-        level: widget.log.level,
-      );
-    }
     setState(() {});
+
+    HCLog.handleLog(
+      type: HCLogType.local,
+      level: widget.log.level,
+      response: response,
+      showLog: widget.log.showLog,
+    );
 
     if (response!.expiredAt <= DateTime.now().millisecondsSinceEpoch) {
       _fetch();
@@ -155,26 +145,11 @@ class _HttpCacheState<T> extends State<HttpCache<T>> {
     try {
       http.Response response = await http.get(
         Uri.parse(url),
-        headers: widget.headers,
+        headers: widget.futureHeaders != null
+            ? await widget.futureHeaders
+            : widget.headers,
       );
 
-      if (widget.log.showLog) {
-        developer.log(
-          response.statusCode.toString(),
-          name: '[server response] Status code',
-          level: widget.log.level,
-        );
-        developer.log(
-          response.headers.toString(),
-          name: '[server response] header',
-          level: widget.log.level,
-        );
-        developer.log(
-          response.body,
-          name: '[server response] body',
-          level: widget.log.level,
-        );
-      }
       this.response = HttpResponse(
         body: response.body,
         statusCode: response.statusCode,
@@ -185,9 +160,14 @@ class _HttpCacheState<T> extends State<HttpCache<T>> {
       );
       await HttpCache.storage.write(url, this.response!.toMap());
 
-      setState(() {
-        isLoading = false;
-      });
+      HCLog.handleLog(
+        type: HCLogType.local,
+        level: widget.log.level,
+        response: this.response,
+        showLog: widget.log.showLog,
+      );
+
+      _setLoading(false);
     } catch (e) {
       if (widget.onError != null) widget.onError!(e);
       error = e;