+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.1-dev.3
+
+- Update doc for pre release
+- Improve log
+- fixing future header
+
 ## 0.0.1-dev.2
 
 - Update doc for pre release